Module kube::api

API helpers

Structs

Api
An easy Api interaction helper

CrBuilder
A builder for CustomResource

CustomResource
A data equivalent of the Resource trait for for Custom Resources

DeleteParams
Common query parameters for delete calls

ListMeta
ListMeta describes metadata that synthetic resources must have, including lists and various status objects. A resource may have only one of {ObjectMeta, ListMeta}.

ListParams
Common query parameters used in watch/list/delete calls on collections

LogParams

NotUsed
Empty struct for when data should be discarded

Object
A standard kubernetes object with .spec and .status

ObjectList
A generic kubernetes object list

ObjectMeta
ObjectMeta is metadata that all persisted resources must have, which includes all objects users must create.

PatchParams
Common query parameters for patch calls

PostParams
Common query parameters for put/post calls

Resource
The Resource information needed to operate a kubernetes client

ScaleSpec
ScaleSpec describes the attributes of a scale subresource.

ScaleStatus
ScaleStatus represents the current status of a scale subresource.

TypeMeta
A convenience struct for ad-hoc serialization

Enums

PatchStrategy
For patch different patch types are supported. See https://kubernetes.io/docs/tasks/run-application/update-api-object-kubectl-patch/#use-a-json-merge-patch-to-update-a-deployment Apply strategy is kinda special

PropagationPolicy
Propagation policy when deleting single objects

WatchEvent
A raw event returned from a watch query

Traits

LoggingObject
Marker trait for objects that has logs

Meta
An accessor trait for Metadata