Struct kube::api::Api

pub struct Api<K> { /* fields omitted */ }

A typed Api variant that does not expose request internals

The upsides of working with this rather than RawApi direct are:

easiers interface (no figuring out return types)

openapi types for free

But the downsides are:

k8s-openapi dependency required (behind feature)

openapi types are unnecessarily heavy on Option use

memory intensive structs because they contain the full data

no control over requests (opinionated)

Methods

impl<K> Api<K>

Expose same interface as Api for controlling scope/group/versions/ns

pub fn within(self, ns: &str) -> Self

pub fn group(self, group: &str) -> Self

pub fn version(self, version: &str) -> Self

impl<K> Api<K> where
K: Clone + DeserializeOwned + KubeObject,

PUSH/PUT/POST/GET abstractions

pub fn get(&self, name: &str) -> Result<K>

pub fn create(&self, pp: &PostParams, data: Vec<u8>) -> Result<K>

pub fn delete(&self, name: &str, dp: &DeleteParams) -> Result<Either<K, Status>>

pub fn list(&self, lp: &ListParams) -> Result<ObjectList<K>>

pub fn delete_collection(
&self,
lp: &ListParams
) -> Result<Either<ObjectList<K>, Status>>

pub fn patch(&self, name: &str, pp: &PatchParams, patch: Vec<u8>) -> Result<K>

pub fn replace(&self, name: &str, pp: &PostParams, data: Vec<u8>) -> Result<K>

pub fn watch(
&self,
lp: &ListParams,
version: &str
) -> Result<Vec<WatchEvent<K>>>

pub fn get_status(&self, name: &str) -> Result<K>

pub fn patch_status(
&self,
name: &str,
pp: &PatchParams,
patch: Vec<u8>
) -> Result<K>

pub fn replace_status(
&self,
name: &str,
pp: &PostParams,
data: Vec<u8>
) -> Result<K>

impl<K> Api<K> where
K: Clone + DeserializeOwned,

Scale subresource

https://kubernetes.io/docs/tasks/access-kubernetes-api/custom-resources/custom-resource-definitions/#scale-subresource

pub fn get_scale(&self, name: &str) -> Result<Scale>

pub fn patch_scale(
&self,
name: &str,
pp: &PatchParams,
patch: Vec<u8>
) -> Result<Scale>

pub fn replace_scale(
&self,
name: &str,
pp: &PostParams,
data: Vec<u8>
) -> Result<Scale>

impl<K> Api<K> where
K: Clone + DeserializeOwned,

Api Constructor for CRDs

Because it relies entirely on user definitions, this ctor does not rely on openapi.

pub fn customResource(client: APIClient, name: &str) -> Self