Struct kube::api::Informer

pub struct Informer<K> where
    K: Clone + DeserializeOwned + KubeObject,  { /* fields omitted */ }

An event informer for a Resource

This watches a Resource<K>, by:

seeding the intial resourceVersion with a list call (optional)

keeping track of resourceVersions after every poll

recovering when resourceVersions get desynced

It caches WatchEvent internally in a queue when polling. A user should drain this queue periodically.

Methods

impl<K> Informer<K> where
K: Clone + DeserializeOwned + KubeObject,

pub fn new(r: Api<K>) -> Self

Create a reflector with a kube client on a kube resource

impl<K> Informer<K> where
K: Clone + DeserializeOwned + KubeObject,

pub fn raw(client: APIClient, r: RawApi) -> Self

Create a reflector with a kube client on a kube resource

pub fn timeout(self, timeout_secs: u32) -> Self

Configure the timeout for the list/watch call.

This limits the duration of the call, regardless of any activity or inactivity. Defaults to 10s

pub fn fields(self, field_selector: &str) -> Self

Configure the selector to restrict the list of returned objects by their fields.

Defaults to everything. Supports '=', '==', and '!=', and can comma separate: key1=value1,key2=value2 The server only supports a limited number of field queries per type.

pub fn labels(self, label_selector: &str) -> Self

Configure the selector to restrict the list of returned objects by their labels.

Defaults to everything. Supports '=', '==', and '!=', and can comma separate: key1=value1,key2=value2

pub fn include_uninitialized(self) -> Self

If called, partially initialized resources are included in watch/list responses.

pub fn init(self) -> Result<Self>

Initialize without a prior version

Will seed resourceVersion with a 1 limit list call to the resource

pub fn init_from(self, v: String) -> Self

Initialize from a prior version

pub fn poll(&self) -> Result<()>

Run a single watch poll

If this returns an error, it resets the resourceVersion. This is meant to be run continually and events are meant to be handled between. If handling all the events is too time consuming, you probably need a queue.

pub fn pop(&self) -> Option<WatchEvent<K>>

Pop an event from the front of the WatchQueue

pub fn reset(&self) -> Result<()>

Reset the resourceVersion to current and clear the event queue

pub fn version(&self) -> String

Return the current version