Struct kube::api::ListParams

pub struct ListParams {
    pub label_selector: Option<String>,
    pub field_selector: Option<String>,
    pub timeout: Option<u32>,
    pub allow_bookmarks: bool,
    pub limit: Option<u32>,
    pub continue_token: Option<String>,
}

Common query parameters used in watch/list/delete calls on collections

Fields
label_selector: Option<String>

A selector to restrict the list of returned objects by their labels.

Defaults to everything if None.

field_selector: Option<String>

A selector to restrict the list of returned objects by their fields.

Defaults to everything if None.

timeout: Option<u32>

Timeout for the list/watch call.

This limits the duration of the call, regardless of any activity or inactivity. If unset for a watch call, we will use 290s. We limit this to 295s due to inherent watch limitations.

allow_bookmarks: bool

Enables watch events with type "BOOKMARK".

Servers that do not implement bookmarks ignore this flag and bookmarks are sent at the server's discretion. Clients should not assume bookmarks are returned at any specific interval, nor may they assume the server will send any BOOKMARK event during a session. If this is not a watch, this field is ignored. If the feature gate WatchBookmarks is not enabled in apiserver, this field is ignored.

limit: Option<u32>

Limit the number of results.

If there are more results, the server will respond with a continue token which can be used to fetch another page of results. See the Kubernetes API docs for pagination details.

continue_token: Option<String>

Fetch a second page of results.

After listing results with a limit, a continue token can be used to fetch another page of results.

Implementations

impl ListParams

Builder interface to ListParams

Usage:

use kube::api::ListParams;
let lp = ListParams::default()
    .timeout(60)
    .labels("kubernetes.io/lifecycle=spot");

pub fn timeout(self, timeout_secs: u32) -> Self

Configure the timeout for list/watch calls

This limits the duration of the call, regardless of any activity or inactivity. Defaults to 290s

pub fn fields(self, field_selector: &str) -> Self

Configure the selector to restrict the list of returned objects by their fields.

Defaults to everything. Supports =, ==, !=, and can be comma separated: key1=value1,key2=value2. The server only supports a limited number of field queries per type.

pub fn labels(self, label_selector: &str) -> Self

Configure the selector to restrict the list of returned objects by their labels.

Defaults to everything. Supports =, ==, !=, and can be comma separated: key1=value1,key2=value2.

pub fn allow_bookmarks(self) -> Self

Enables watch bookmarks from the api server if supported

pub fn limit(self, limit: u32) -> Self

Sets a result limit.

pub fn continue_token(self, token: &str) -> Self

Sets a continue token.