Struct kube::api::RawApi

pub struct RawApi {
    pub resource: String,
    pub group: String,
    pub namespace: Option<String>,
    pub version: String,
    pub prefix: String,
}

RawRawApi generation data

This data defines the urls used by kubernetes' APIs. This struct is client agnostic, and can be passed to an Informer or a Reflector.

Can be used directly with a client. When data is PUT/POST/PATCH'd this struct requires serialized raw bytes.

Fields
resource: String

API Resource name

group: String

API Group

namespace: Option<String>

Namespace the resources reside

version: String

API version of the resource

prefix: String

Name of the api prefix (api or apis typically)

Methods

impl RawApi

Constructors for most kubernetes objects

Don't see all objects in here? Please submit a PR. You can extract the data needed from the openapi spec.

pub fn within(self, ns: &str) -> Self

Set as namespaced resource within a specified namespace

pub fn group(self, group: &str) -> Self

Set the api group of a resource manually

Can be used to set legacy versions like "extensions" for old Deployments

pub fn version(self, version: &str) -> Self

Set the version of an api group manually

Can be used to set legacy versions like "v1beta1" for old Deployments

pub fn v1Namespace() -> Self

Stable namespace resource constructor

pub fn v1Deployment() -> Self

Stable deployment resource constructor

pub fn v1Pod() -> Self

Stable pod resource constructor

pub fn v1DaemonSet() -> Self

Stable daemonset resource constructor

pub fn v1ReplicaSet() -> Self

Stable replicaset resource constructor

pub fn v1Node() -> Self

Stable node resource constructor

pub fn v1Statefulset() -> Self

Stable statefulset resource constructor

pub fn v1Event() -> Self

pub fn v1Service() -> Self

pub fn v1Secret() -> Self

pub fn v1ConfigMap() -> Self

pub fn v1beta1CustomResourceDefinition() -> Self

Custom resource definition constructor

pub fn customResource(name: &str) -> Self

Instance of a CRD

The version, and group must be set by the user:

use kube::api::RawApi;
let foos = RawApi::customResource("foos") // <.spec.name>
   .group("clux.dev") // <.spec.group>
   .version("v1");

impl RawApi

Convenience methods found from API conventions

pub fn list(&self, lp: &ListParams) -> Result<Request<Vec<u8>>>

List a collection of a resource

pub fn watch(&self, lp: &ListParams, ver: &str) -> Result<Request<Vec<u8>>>

Watch a resource at a given version

pub fn get(&self, name: &str) -> Result<Request<Vec<u8>>>

Get a single instance

pub fn create(&self, pp: &PostParams, data: Vec<u8>) -> Result<Request<Vec<u8>>>

Create an instance of a resource

pub fn delete(&self, name: &str, dp: &DeleteParams) -> Result<Request<Vec<u8>>>

Delete an instance of a resource

pub fn delete_collection(&self, lp: &ListParams) -> Result<Request<Vec<u8>>>

Delete a collection of a resource

pub fn patch(
&self,
name: &str,
pp: &PatchParams,
patch: Vec<u8>
) -> Result<Request<Vec<u8>>>

Patch an instance of a resource

Requires a serialized merge-patch+json at the moment.

pub fn replace(
&self,
name: &str,
pp: &PostParams,
data: Vec<u8>
) -> Result<Request<Vec<u8>>>

Replace an instance of a resource

Requires metadata.resourceVersion set in data

pub fn get_scale(&self, name: &str) -> Result<Request<Vec<u8>>>

Get an instance of the scale subresource

pub fn patch_scale(
&self,
name: &str,
pp: &PatchParams,
patch: Vec<u8>
) -> Result<Request<Vec<u8>>>

Patch an instance of the scale subresource

pub fn replace_scale(
&self,
name: &str,
pp: &PostParams,
data: Vec<u8>
) -> Result<Request<Vec<u8>>>

Replace an instance of the scale subresource

pub fn get_status(&self, name: &str) -> Result<Request<Vec<u8>>>

Get an instance of the status subresource

pub fn patch_status(
&self,
name: &str,
pp: &PatchParams,
patch: Vec<u8>
) -> Result<Request<Vec<u8>>>

Patch an instance of the status subresource

pub fn replace_status(
&self,
name: &str,
pp: &PostParams,
data: Vec<u8>
) -> Result<Request<Vec<u8>>>

Replace an instance of the status subresource

impl RawApi

pub fn log(&self, name: &str, lp: &LogParams) -> Result<Request<Vec<u8>>>

Get a pod logs