Struct kube::api::Resource

pub struct Resource {
    pub api_version: String,
    pub group: String,
    pub kind: String,
    pub version: String,
    pub namespace: Option<String>,
}

A Kubernetes resource that can be accessed through the API

Fields
api_version: String

The API version of the resource.

This is a composite of Resource::GROUP and Resource::VERSION (eg “apiextensions.k8s.io/v1beta1”) or just the version for resources without a group (eg “v1”). This is the string used in the apiVersion field of the resource’s serialized form.

group: String

The group of the resource

or the empty string if the resource doesn’t have a group.

kind: String

The kind of the resource.

This is the string used in the kind field of the resource’s serialized form.

version: String

The version of the resource.

namespace: Option<String>

The namespace if the resource resides (if namespaced)

Implementations

impl Resource

pub fn all<K: Resource>() -> Self

Cluster level resources, or resources viewed across all namespaces

pub fn namespaced<K: Resource>(ns: &str) -> Self

Namespaced resource within a given namespace

pub fn dynamic(kind: &str) -> DynamicResource

Manually configured resource or custom resource

This is the only entrypoint to Resource that bypasses k8s_openapi entirely. If you need a CustomResource consider using kube-derive for its #[derive(CustomResource)] proc-macro.

impl Resource

Convenience methods found from API conventions

pub fn list(&self, lp: &ListParams) -> Result<Request<Vec<u8>>>

List a collection of a resource

pub fn watch(&self, lp: &ListParams, ver: &str) -> Result<Request<Vec<u8>>>

Watch a resource at a given version

pub fn get(&self, name: &str) -> Result<Request<Vec<u8>>>

Get a single instance

pub fn create(&self, pp: &PostParams, data: Vec<u8>) -> Result<Request<Vec<u8>>>

Create an instance of a resource

pub fn delete(&self, name: &str, dp: &DeleteParams) -> Result<Request<Vec<u8>>>

Delete an instance of a resource

pub fn delete_collection(
&self,
dp: &DeleteParams,
lp: &ListParams
) -> Result<Request<Vec<u8>>>

Delete a collection of a resource

pub fn patch<P: Serialize>(
&self,
name: &str,
pp: &PatchParams,
patch: &Patch<P>
) -> Result<Request<Vec<u8>>>

Patch an instance of a resource

Requires a serialized merge-patch+json at the moment.

pub fn replace(
&self,
name: &str,
pp: &PostParams,
data: Vec<u8>
) -> Result<Request<Vec<u8>>>

Replace an instance of a resource

Requires metadata.resourceVersion set in data

impl Resource

Scale subresource

pub fn get_scale(&self, name: &str) -> Result<Request<Vec<u8>>>

Get an instance of the scale subresource

pub fn patch_scale<P: Serialize>(
&self,
name: &str,
pp: &PatchParams,
patch: &Patch<P>
) -> Result<Request<Vec<u8>>>

Patch an instance of the scale subresource

pub fn replace_scale(
&self,
name: &str,
pp: &PostParams,
data: Vec<u8>
) -> Result<Request<Vec<u8>>>

Replace an instance of the scale subresource

impl Resource

Status subresource

pub fn get_status(&self, name: &str) -> Result<Request<Vec<u8>>>

Get an instance of the status subresource

pub fn patch_status<P: Serialize>(
&self,
name: &str,
pp: &PatchParams,
patch: &Patch<P>
) -> Result<Request<Vec<u8>>>

Patch an instance of the status subresource

pub fn replace_status(
&self,
name: &str,
pp: &PostParams,
data: Vec<u8>
) -> Result<Request<Vec<u8>>>

Replace an instance of the status subresource

impl Resource

pub fn logs(&self, name: &str, lp: &LogParams) -> Result<Request<Vec<u8>>>

Get a pod logs

impl Resource

pub fn evict(&self, name: &str, ep: &EvictParams) -> Result<Request<Vec<u8>>>

Create an eviction