Trait kube::api::Meta

pub trait Meta: Metadata {
    fn meta(&self) -> &ObjectMeta;

    fn name(&self) -> String;

    fn namespace(&self) -> Option<String>;

    fn resource_ver(&self) -> Option<String>;
}

An accessor trait for Metadata

This for a subset of kubernetes type that do not end in List These types, using ObjectMeta, SHOULD all have required properties:

.metadata

.metadata.name And these optional properties:

.metadata.namespace

.metadata.resource_version

This avoids a bunch of the unnecessary unwrap mechanics for apps

Required methods

fn meta(&self) -> &ObjectMeta

fn name(&self) -> String

fn namespace(&self) -> Option<String>

fn resource_ver(&self) -> Option<String>

Loading content...