kube
Module client

A basic API client for interacting with the Kubernetes API

The Client uses standard kube error handling.

This client can be used on its own or in conjuction with the Api type for more structured interaction with the kubernetes API.

The Client can also be used with Discovery to dynamically retrieve the resources served by the kubernetes API.

Modules

middleware

Middleware types returned from ConfigExt methods.

Structs

Client

Client for connecting with a Kubernetes cluster.

ClientBuilder

Builder for Client instances with customized tower middleware.

Status

A Kubernetes status object

Enums

AuthError

Client auth errors

RustlsTlsError

Errors from Rustls

Traits

ConfigExt

Extensions to Config for custom Client.

Type Aliases

DynBody

HTTP body of a dynamic backing type.