Struct kube::client::Client

pub struct Client { /* fields omitted */ }

Client for connecting with a Kubernetes cluster.

The best way to instantiate the client is either by inferring the configuration from the environment using Client::try_default or with an existing Config using Client::try_from.

Implementations

impl Client

pub fn new(service: Service) -> Self

Create and initialize a Client using the given Service.

pub async fn try_default() -> Result<Self>

Create and initialize a Client using the inferred configuration.

Will use Config::infer to try in-cluster enironment variables first, then fallback to the local kubeconfig.

Will fail if neither configuration could be loaded.

If you already have a Config then use Client::try_from instead

pub async fn request<T>(&self, request: Request<Vec<u8>>) -> Result<T> where
T: DeserializeOwned,

Perform a raw HTTP request against the API and deserialize the response as JSON to some known type.

pub async fn request_text(&self, request: Request<Vec<u8>>) -> Result<String>

Perform a raw HTTP request against the API and get back the response as a string

pub async fn request_text_stream(
&self,
request: Request<Vec<u8>>
) -> Result<impl Stream<Item = Result<Bytes>>>

Perform a raw HTTP request against the API and get back the response as a stream of bytes

pub async fn request_status<T>(
&self,
request: Request<Vec<u8>>
) -> Result<Either<T, Status>> where
T: DeserializeOwned,

Perform a raw HTTP request against the API and get back either an object deserialized as JSON or a Status Object.

pub async fn request_events<T: Clone + Meta>(
&self,
request: Request<Vec<u8>>
) -> Result<impl TryStream<Item = Result<WatchEvent<T>>>> where
T: DeserializeOwned,

Perform a raw request and get back a stream of WatchEvent objects

pub async fn apiserver_version(&self) -> Result<Info>

Returns apiserver version.

pub async fn list_api_groups(&self) -> Result<APIGroupList>

Lists api groups that apiserver serves.

pub async fn list_api_group_resources(
&self,
apiversion: &str
) -> Result<APIResourceList>

Lists resources served in given API group.

Example usage:

let apigroups = client.list_api_groups().await?;
for g in apigroups.groups {
    let ver = g
        .preferred_version
        .as_ref()
        .or_else(|| g.versions.first())
        .expect("preferred or versions exists");
    let apis = client.list_api_group_resources(&ver.group_version).await?;
    dbg!(apis);
}

pub async fn list_core_api_versions(&self) -> Result<APIVersions>

Lists versions of core a.k.a. "" legacy API group.

pub async fn list_core_api_resources(
&self,
version: &str
) -> Result<APIResourceList>

Lists resources served in particular core group version.