Struct kube::client::ClientBuilder

pub struct ClientBuilder<Svc> { /* private fields */ }

Builder for Client instances with customized tower middleware.

Implementations

impl<Svc> ClientBuilder<Svc>

pub fn new( service: Svc, default_namespace: impl Into<String> ) -> ClientBuilder<Svc>where Svc: Service<Request<Body>>,

Construct a ClientBuilder from scratch with a fully custom Service stack.

This method is only intended for advanced use cases, most users will want to use ClientBuilder::try_from instead, which provides a default stack as a starting point.

pub fn with_layer<L>( self, layer: &L ) -> ClientBuilder<<L as Layer<Svc>>::Service>where L: Layer<Svc>,

Add a Layer to the current Service stack.

pub fn build<B>(self) -> Clientwhere Svc: Service<Request<Body>, Response = Response<B>> + Send + 'static, <Svc as Service<Request<Body>>>::Future: Send + 'static, <Svc as Service<Request<Body>>>::Error: Into<Box<dyn Error + Sync + Send + 'static, Global>>, B: Body<Data = Bytes> + Send + 'static, <B as Body>::Error: Into<Box<dyn Error + Sync + Send + 'static, Global>>,

Build a Client instance with the current Service stack.