Module kube::config

In cluster or out of cluster kubeconfig to be used by an api client

You primarily want to interact with Configuration, and its associated load functions.

The full Config and child-objects are exposed here for convenience only.

Structs

AuthInfo
AuthInfo stores information to tell cluster who you are.

AuthProviderConfig
AuthProviderConfig stores auth for specified cloud provider.

Cluster
Cluster stores information to connect kubernetes cluster.

Config
Config stores information to connect remote kubernetes cluster.

ConfigOptions
ConfigOptions stores options used when loading kubeconfig file.

Configuration
Configuration stores kubernetes path and client for requests.

Context
Context stores tuple of cluster and user information.

ExecConfig
ExecConfig stores credential-plugin configuration.

NamedCluster
NamedCluster associates name with cluster.

NamedContext
NamedContext associates name with context.

NamedExtension
NamedExtention associates name with extension.

Preferences
Preferences stores extensions for cli.

Functions

create_client_builder
Returns a client builder and config loader, based on the cluster information from the kubeconfig file.

incluster_config
Returns a config which is used by clients within pods on kubernetes.

load_kube_config
Returns a config includes authentication and cluster infomation from kubeconfig file.

load_kube_config_with
Returns a config which includes authentication and cluster information from kubeconfig file.