kube::core::discovery

Module v2

Types for the Aggregated Discovery API (apidiscovery.k8s.io/v2)

These types are not part of the Kubernetes OpenAPI spec, so they are defined here rather than in k8s-openapi. They mirror the types from k8s.io/api/apidiscovery/v2.

The Aggregated Discovery API is available since Kubernetes 1.26 (beta) and stable in 1.30+.

Structs

APIGroupDiscovery

APIGroupDiscovery holds information about which resources are being served for all version of the API Group. It contains a list of APIVersionDiscovery that holds a list of APIResourceDiscovery types served for a version. Versions are in descending order of preference, with the first version being the preferred entry.

APIGroupDiscoveryList

APIGroupDiscoveryList is a resource containing a list of APIGroupDiscovery. This is one of the types that can be returned from the /api and /apis endpoint and contains an aggregated list of API resources (built-ins, Custom Resource Definitions, resources from aggregated servers) that a cluster supports.

APIResourceDiscovery

APIResourceDiscovery provides information about an API resource for discovery.

APISubresourceDiscovery

APISubresourceDiscovery provides information about an API subresource for discovery.

APIVersionDiscovery

APIVersionDiscovery holds a list of APIResourceDiscovery types that are served for a particular version within an API Group.

GroupVersionKind

GroupVersionKind unambiguously identifies a kind. This is a local copy for use in discovery types.

Constants

ACCEPT_AGGREGATED_DISCOVERY_V2

Content negotiation Accept header for Aggregated Discovery API v2