Module kube::core::metadata

Metadata structs used in traits, lists, and dynamic objects.

Structs

ListMeta

ListMeta describes metadata that synthetic resources must have, including lists and various status objects. A resource may have only one of {ObjectMeta, ListMeta}.

ObjectMeta

ObjectMeta is metadata that all persisted resources must have, which includes all objects users must create.

PartialObjectMeta

A generic representation of any object with ObjectMeta.

TypeMeta

Type information that is flattened into every kubernetes object

Traits

PartialObjectMetaExt

Helper trait for converting ObjectMeta into useful PartialObjectMeta variants