kube::core
Module params

A port of request parameter *Optionals from apimachinery/types.go

Structs

DeleteParams

Common query parameters for delete calls

GetParams

Common query parameters used in get calls

ListParams

Common query parameters used in list/delete calls on collections

PatchParams

Common query parameters for patch calls

PostParams

Common query parameters for put/post calls

Preconditions

Preconditions must be fulfilled before an operation (update, delete, etc.) is carried out.

WatchParams

Common query parameters used in watch calls on collections

Enums

Patch

Describes changes that should be applied to a resource

PropagationPolicy

Propagation policy when deleting single objects

ValidationDirective

The validation directive to use for fieldValidation when using server-side apply.

VersionMatch

Controls how the resource version parameter is applied for list calls