Struct kube::core::request::Request

pub struct Request {
    pub url_path: String,
}

A Kubernetes request builder

Takes a base_path and supplies constructors for common operations The extra operations all return http::Request objects.

Fields
url_path: String

The path component of a url

Implementations

impl Request

pub fn new<S>(url_path: S) -> Request
where S: Into<String>,

New request with a resource’s url path

impl Request

Convenience methods found from API conventions

pub fn list(&self, lp: &ListParams) -> Result<Request<Vec<u8>>, Error>

List a collection of a resource

pub fn watch( &self, wp: &WatchParams, ver: &str, ) -> Result<Request<Vec<u8>>, Error>

Watch a resource at a given version

pub fn get(&self, name: &str, gp: &GetParams) -> Result<Request<Vec<u8>>, Error>

Get a single instance

pub fn create( &self, pp: &PostParams, data: Vec<u8>, ) -> Result<Request<Vec<u8>>, Error>

Create an instance of a resource

pub fn delete( &self, name: &str, dp: &DeleteParams, ) -> Result<Request<Vec<u8>>, Error>

Delete an instance of a resource

pub fn delete_collection( &self, dp: &DeleteParams, lp: &ListParams, ) -> Result<Request<Vec<u8>>, Error>

Delete a collection of a resource

pub fn patch<P>( &self, name: &str, pp: &PatchParams, patch: &Patch<P>, ) -> Result<Request<Vec<u8>>, Error>
where P: Serialize,

Patch an instance of a resource

Requires a serialized merge-patch+json at the moment.

pub fn replace( &self, name: &str, pp: &PostParams, data: Vec<u8>, ) -> Result<Request<Vec<u8>>, Error>

Replace an instance of a resource

Requires metadata.resourceVersion set in data

impl Request

Subresources

pub fn get_subresource( &self, subresource_name: &str, name: &str, ) -> Result<Request<Vec<u8>>, Error>

Get an instance of the subresource

pub fn create_subresource( &self, subresource_name: &str, name: &str, pp: &PostParams, data: Vec<u8>, ) -> Result<Request<Vec<u8>>, Error>

Create an instance of the subresource

pub fn patch_subresource<P>( &self, subresource_name: &str, name: &str, pp: &PatchParams, patch: &Patch<P>, ) -> Result<Request<Vec<u8>>, Error>
where P: Serialize,

Patch an instance of the subresource

pub fn replace_subresource( &self, subresource_name: &str, name: &str, pp: &PostParams, data: Vec<u8>, ) -> Result<Request<Vec<u8>>, Error>

Replace an instance of the subresource

impl Request

Metadata-only request implementations

Requests set an extended Accept header compromised of JSON media type and additional parameters that retrieve only necessary metadata from an object.

pub fn get_metadata( &self, name: &str, gp: &GetParams, ) -> Result<Request<Vec<u8>>, Error>

Get a single metadata instance for a named resource

pub fn list_metadata(&self, lp: &ListParams) -> Result<Request<Vec<u8>>, Error>

List a collection of metadata of a resource

pub fn watch_metadata( &self, wp: &WatchParams, ver: &str, ) -> Result<Request<Vec<u8>>, Error>

Watch metadata of a resource at a given version

pub fn patch_metadata<P>( &self, name: &str, pp: &PatchParams, patch: &Patch<P>, ) -> Result<Request<Vec<u8>>, Error>
where P: Serialize,

Patch an instance of a resource and receive its metadata only

Requires a serialized merge-patch+json at the moment

impl Request

pub fn logs( &self, name: &str, lp: &LogParams, ) -> Result<Request<Vec<u8>>, Error>

Get a pod logs

impl Request

pub fn evict( &self, name: &str, ep: &EvictParams, ) -> Result<Request<Vec<u8>>, Error>

Create an eviction

impl Request

pub fn restart(&self, name: &str) -> Result<Request<Vec<u8>>, Error>

Restart a resource

impl Request

pub fn cordon(&self, name: &str) -> Result<Request<Vec<u8>>, Error>

Cordon a resource

pub fn uncordon(&self, name: &str) -> Result<Request<Vec<u8>>, Error>

Uncordon a resource