kube::core

Struct DeserializeGuard

pub struct DeserializeGuard<K>(pub Result<K, InvalidObject>);

A wrapper type for K that lets deserializing the parent object succeed, even if the K is invalid.

For example, this can be used to still access valid objects from an Api::list call or watcher.

Tuple Fields
0: Result<K, InvalidObject>