Struct kube::core::subresource::Scale

pub struct Scale {
    pub metadata: ObjectMeta,
    pub spec: Option<ScaleSpec>,
    pub status: Option<ScaleStatus>,
}

Scale represents a scaling request for a resource.

Fields
metadata: ObjectMeta

Standard object metadata; More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata.

spec: Option<ScaleSpec>

defines the behavior of the scale. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status.

status: Option<ScaleStatus>

current status of the scale. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status. Read-only.

Implementations

impl Scale

pub fn patch_deployment( name: &str, namespace: &str, body: &Patch, optional: PatchOptional<'_> ) -> Result<(Request<Vec<u8, Global>>, fn(_: StatusCode) -> ResponseBody<PatchResponse<Scale>>), RequestError>

partially update scale of the specified Deployment

Use the returned crate::ResponseBody<crate::PatchResponse<Self>> constructor, or crate::PatchResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Scale

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Scale

pub fn patch_replica_set( name: &str, namespace: &str, body: &Patch, optional: PatchOptional<'_> ) -> Result<(Request<Vec<u8, Global>>, fn(_: StatusCode) -> ResponseBody<PatchResponse<Scale>>), RequestError>

partially update scale of the specified ReplicaSet

Use the returned crate::ResponseBody<crate::PatchResponse<Self>> constructor, or crate::PatchResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Scale

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Scale

pub fn patch_stateful_set( name: &str, namespace: &str, body: &Patch, optional: PatchOptional<'_> ) -> Result<(Request<Vec<u8, Global>>, fn(_: StatusCode) -> ResponseBody<PatchResponse<Scale>>), RequestError>

partially update scale of the specified StatefulSet

Use the returned crate::ResponseBody<crate::PatchResponse<Self>> constructor, or crate::PatchResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Scale

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Scale

pub fn patch_replication_controller( name: &str, namespace: &str, body: &Patch, optional: PatchOptional<'_> ) -> Result<(Request<Vec<u8, Global>>, fn(_: StatusCode) -> ResponseBody<PatchResponse<Scale>>), RequestError>

partially update scale of the specified ReplicationController

Use the returned crate::ResponseBody<crate::PatchResponse<Self>> constructor, or crate::PatchResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Scale

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Scale

pub fn read_deployment( name: &str, namespace: &str ) -> Result<(Request<Vec<u8, Global>>, fn(_: StatusCode) -> ResponseBody<ReadDeploymentScaleResponse>), RequestError>

read scale of the specified Deployment

Use the returned crate::ResponseBody<ReadDeploymentScaleResponse> constructor, or ReadDeploymentScaleResponse directly, to parse the HTTP response.

Arguments

name

name of the Scale

namespace

object name and auth scope, such as for teams and projects

impl Scale

pub fn read_replica_set( name: &str, namespace: &str ) -> Result<(Request<Vec<u8, Global>>, fn(_: StatusCode) -> ResponseBody<ReadReplicaSetScaleResponse>), RequestError>

read scale of the specified ReplicaSet

Use the returned crate::ResponseBody<ReadReplicaSetScaleResponse> constructor, or ReadReplicaSetScaleResponse directly, to parse the HTTP response.

Arguments

name

name of the Scale

namespace

object name and auth scope, such as for teams and projects

impl Scale

pub fn read_stateful_set( name: &str, namespace: &str ) -> Result<(Request<Vec<u8, Global>>, fn(_: StatusCode) -> ResponseBody<ReadStatefulSetScaleResponse>), RequestError>

read scale of the specified StatefulSet

Use the returned crate::ResponseBody<ReadStatefulSetScaleResponse> constructor, or ReadStatefulSetScaleResponse directly, to parse the HTTP response.

Arguments

name

name of the Scale

namespace

object name and auth scope, such as for teams and projects

impl Scale

pub fn read_replication_controller( name: &str, namespace: &str ) -> Result<(Request<Vec<u8, Global>>, fn(_: StatusCode) -> ResponseBody<ReadReplicationControllerScaleResponse>), RequestError>

read scale of the specified ReplicationController

Use the returned crate::ResponseBody<ReadReplicationControllerScaleResponse> constructor, or ReadReplicationControllerScaleResponse directly, to parse the HTTP response.

Arguments

name

name of the Scale

namespace

object name and auth scope, such as for teams and projects

impl Scale

pub fn replace_deployment( name: &str, namespace: &str, body: &Scale, optional: ReplaceOptional<'_> ) -> Result<(Request<Vec<u8, Global>>, fn(_: StatusCode) -> ResponseBody<ReplaceResponse<Scale>>), RequestError>

replace scale of the specified Deployment

Use the returned crate::ResponseBody<crate::ReplaceResponse<Self>> constructor, or crate::ReplaceResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Scale

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Scale

pub fn replace_replica_set( name: &str, namespace: &str, body: &Scale, optional: ReplaceOptional<'_> ) -> Result<(Request<Vec<u8, Global>>, fn(_: StatusCode) -> ResponseBody<ReplaceResponse<Scale>>), RequestError>

replace scale of the specified ReplicaSet

Use the returned crate::ResponseBody<crate::ReplaceResponse<Self>> constructor, or crate::ReplaceResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Scale

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Scale

pub fn replace_stateful_set( name: &str, namespace: &str, body: &Scale, optional: ReplaceOptional<'_> ) -> Result<(Request<Vec<u8, Global>>, fn(_: StatusCode) -> ResponseBody<ReplaceResponse<Scale>>), RequestError>

replace scale of the specified StatefulSet

Use the returned crate::ResponseBody<crate::ReplaceResponse<Self>> constructor, or crate::ReplaceResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Scale

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Scale

pub fn replace_replication_controller( name: &str, namespace: &str, body: &Scale, optional: ReplaceOptional<'_> ) -> Result<(Request<Vec<u8, Global>>, fn(_: StatusCode) -> ResponseBody<ReplaceResponse<Scale>>), RequestError>

replace scale of the specified ReplicationController

Use the returned crate::ResponseBody<crate::ReplaceResponse<Self>> constructor, or crate::ReplaceResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Scale

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.