kube

Module discovery

High-level utilities for runtime API discovery.

Modules

oneshot

single use discovery utils

verbs

Rbac verbs for ApiCapabilities

Structs

ApiCapabilities

Contains the capabilities of an API resource

ApiGroup

Describes one API groups collected resources and capabilities.

ApiResource

Information about a Kubernetes API resource

Discovery

A caching client for running API discovery against the Kubernetes API.

Enums

Scope

Resource scope

Functions

group

Discovers all APIs available under a certain group at all versions

pinned_group

Discovers all APIs available under a certain group at a pinned version

pinned_kind

Single discovery for a single GVK