Module kube::discovery::oneshot

single use discovery utils

These helpers provides a simpler discovery interface, but do not offer any built-in caching.

This can provide specific information for 3 cases:

single kind in a particular group at a pinned version via oneshot::pinned_kind

all kinds in a group at pinned version: “apiregistration.k8s.io/v1” via oneshot::pinned_group

all kinds/version combinations in a group: “apiregistration.k8s.io” via oneshot::group

Functions

group

Discovers all APIs available under a certain group at all versions

pinned_group

Discovers all APIs available under a certain group at a pinned version

pinned_kind

Single discovery for a single GVK