kube

Module error

Error handling and error types

Structs

Status

A Kubernetes status object

Enums

DiscoveryError

Possible errors when using API discovery

Error

Possible errors from the Client

Type Aliases

ErrorResponseDeprecated

This type used to be a payload for kube::Error::Api. It has been replaced by Status. In the interest of backward compatibility, we keep this type definition here as a deprecated alias. Status better reflects the Kubernetes API conventions and is more versatile. If you need to migrate your code, simply replace ErrorResponse with Status and also check helper methods implemented on Status for easily identifying common error cases.