kube

Crate runtime

Common components for building Kubernetes operators

This crate contains the core building blocks to allow users to build controllers/operators/watchers that need to synchronize/reconcile kubernetes state.

Newcomers are recommended to start with the Controller builder, which gives an opinionated starting point that should be appropriate for simple operators, but all components are designed to be usable á la carte if your operator doesn’t quite fit that mold.

Modules

controller

Runs a user-supplied reconciler function on objects when they (or related objects) are updated

events

Publishes events for objects for kubernetes >= 1.19

finalizer

Finalizer helper for Controller reconcilers

reflector

Caches objects in memory

scheduler

Delays and deduplicates Stream items

utils

Helpers for manipulating built-in streams

wait

Waits for objects to reach desired states

watcher

Watches a Kubernetes Resource for changes, with error recovery

Structs

Config

Accumulates all options that can be used on a Controller invocation.

Controller

Controller for a Resource K

Traits

Predicate

A predicate is a hasher of Kubernetes objects stream filtering

WatchStreamExt

Extension trait for streams returned by watcher or reflector

Functions

applier

Apply a reconciler to an input stream, with a given retry policy

finalizer

Reconcile an object in a way that requires cleanup before an object can be deleted.

metadata_watcher

Watches a Kubernetes Resource for changes continuously and receives only the metadata

reflector

Cache objects from a watcher() stream into a local Store

scheduler

Stream transformer that delays and deduplicates items.

watcher

Watches a Kubernetes Resource for changes continuously