kube::runtime::reflector

Struct ObjectRef

#[non_exhaustive]
pub struct ObjectRef<K>
where
    K: Lookup + ?Sized,
{
    pub dyntype: <K as Lookup>::DynamicType,
    pub name: String,
    pub namespace: Option<String>,
    pub extra: Extra,
}

A typed and namedspaced (if relevant) reference to a Kubernetes object

K may be either the object type or DynamicObject, in which case the type is stored at runtime. Erased ObjectRefs pointing to different types are still considered different.

use kube_runtime::reflector::ObjectRef;
use k8s_openapi::api::core::v1::{ConfigMap, Secret};
assert_ne!(
    ObjectRef::<ConfigMap>::new("a").erase(),
    ObjectRef::<Secret>::new("a").erase(),
);

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

dyntype: <K as Lookup>::DynamicTypename: String

The name of the object

namespace: Option<String>

The namespace of the object

May only be None if the kind is cluster-scoped (not located in a namespace). Note that it is acceptable for an ObjectRef to a cluster-scoped resource to have a namespace. These are, however, not considered equal:

assert_ne!(ObjectRef::<ConfigMap>::new("foo"), ObjectRef::new("foo").within("bar"));

extra: Extra

Extra information about the object being referred to

This is not considered when comparing objects, but may be used when converting to and from other representations, such as OwnerReference or ObjectReference.

Implementations

impl<K> ObjectRef<K>
where K: Lookup, <K as Lookup>::DynamicType: Default,

pub fn new(name: &str) -> ObjectRef<K>

pub fn from_obj(obj: &K) -> ObjectRef<K>

impl<K> ObjectRef<K>
where K: Lookup,

pub fn new_with(name: &str, dyntype: <K as Lookup>::DynamicType) -> ObjectRef<K>

pub fn within(self, namespace: &str) -> ObjectRef<K>

pub fn from_obj_with( obj: &K, dyntype: <K as Lookup>::DynamicType, ) -> ObjectRef<K>
where K: Lookup,

Creates ObjectRef from the resource and dynamic type.

pub fn from_owner_ref( namespace: Option<&str>, owner: &OwnerReference, dyntype: <K as Lookup>::DynamicType, ) -> Option<ObjectRef<K>>

Create an ObjectRef from an OwnerReference

Returns None if the types do not match.

pub fn into_kind_unchecked<K2>( self, dt2: <K2 as Lookup>::DynamicType, ) -> ObjectRef<K2>
where K2: Lookup,

Convert into a reference to K2

Note that no checking is done on whether this conversion makes sense. For example, every Service has a corresponding Endpoints, but it wouldn’t make sense to convert a Pod into a Deployment.

pub fn erase(self) -> ObjectRef<DynamicObject>