kube::runtime

Module utils

Helpers for manipulating built-in streams

Modules

predicates

Predicate functions for WatchStreamExt::predicate_filter

Structs

CancelableJoinHandle

A JoinHandle that cancels the Future when dropped, rather than detaching it

EventDecode

Stream returned by the applied_objects and touched_objects method.

EventFlatten

Stream returned by the applied_objects and touched_objects method.

EventModify

Stream returned by the modify method. Modifies the Event item returned by the inner stream by calling modify on it.

PredicateFilter

Stream returned by the predicate_filter method.

Reflect

Stream returned by the reflect method

ResetTimerBackoff

A Backoff wrapper that resets after a fixed duration has elapsed.

StreamBackoff

Applies a Backoff policy to a Stream

Traits

Backoff

Predicate

A predicate is a hasher of Kubernetes objects stream filtering

WatchStreamExt

Extension trait for streams returned by watcher or reflector