kube::runtime::utils

Module predicates

Predicate functions for WatchStreamExt::predicate_filter

These functions just return a hash of commonly compared values, to help decide whether to pass a watch event along or not.

Functional rewrite of the controller-runtime/predicate module.

Functions

annotations

Hash the annotations of a Resource K

finalizers

Hash the finalizers of a Resource K

generation

Hash the generation of a Resource K

labels

Hash the labels of a Resource K

resource_version

Hash the resource version of a Resource K