kube::runtime

Module wait

Waits for objects to reach desired states

Modules

conditions

Common conditions to wait for

delete

Utilities for deleting objects

Enums

Error

Traits

Condition

A trait for condition functions to be used by await_condition

Functions

await_condition

Watch an object, and wait for some condition cond to return true.