kube::runtime

Module watcher

Watches a Kubernetes Resource for changes, with error recovery

See watcher for the primary entry point.

Structs

Config

Accumulates all options that can be used on the watcher invocation.

DefaultBackoff

Default watcher backoff inspired by Kubernetes’ client-go.

ExponentialBackoff

Enums

Error

Event

Watch events returned from the watcher

InitialListStrategy

Configurable watcher listwatch semantics

ListSemantic

Configurable list semantics for watcher relists

Functions

metadata_watcher

Watches a Kubernetes Resource for changes continuously and receives only the metadata

watch_object

Watch a single named object for updates

watcher

Watches a Kubernetes Resource for changes continuously

Type Aliases

Result