Struct kube::Client

pub struct Client { /* fields omitted */ }

Client for connecting with a Kubernetes cluster.

The easiest way to instantiate the client is either by inferring the configuration from the environment using Client::try_default or with an existing Config using Client::try_from.

Implementations

impl Client

pub fn new<S, B, T>(service: S, default_namespace: T) -> Client where
S: 'static + Service<Request<Body>, Response = Response<B>> + Send,
B: 'static + Body<Data = Bytes> + Send,
T: Into<String>,
<S as Service<Request<Body>>>::Future: 'static,
<S as Service<Request<Body>>>::Future: Send,
<S as Service<Request<Body>>>::Error: Into<Box<dyn Error + Sync + Send + 'static, Global>>,
<B as Body>::Error: 'static,
<B as Body>::Error: Error,
<B as Body>::Error: Send,
<B as Body>::Error: Sync,

Create a Client using a custom Service stack.

ConfigExt provides extensions for building a custom stack.

To create with the default stack with a Config, use Client::try_from.

To create with the default stack with an inferred Config, use Client::try_default.

Example

use kube::{client::ConfigExt, Client, Config};
use tower::ServiceBuilder;

let config = Config::infer().await?;
let service = ServiceBuilder::new()
    .layer(config.base_uri_layer())
    .option_layer(config.auth_layer()?)
    .service(hyper::Client::new());
let client = Client::new(service, config.default_namespace);

pub async fn try_default() -> Result<Client, Error>

Create and initialize a Client using the inferred configuration.

Will use Config::infer to try in-cluster enironment variables first, then fallback to the local kubeconfig.

Will fail if neither configuration could be loaded.

If you already have a Config then use Client::try_from instead.

pub async fn request<T>(
&'_ self,
request: Request<Vec<u8, Global>>
) -> Result<T, Error> where
T: DeserializeOwned,

Perform a raw HTTP request against the API and deserialize the response as JSON to some known type.

pub async fn request_text(
&'_ self,
request: Request<Vec<u8, Global>>
) -> Result<String, Error>

Perform a raw HTTP request against the API and get back the response as a string

pub async fn request_text_stream(
&'_ self,
request: Request<Vec<u8, Global>>
) -> Result<impl Stream<Item = Result<Bytes, Error>>, Error>

Perform a raw HTTP request against the API and get back the response as a stream of bytes

pub async fn request_status<T>(
&'_ self,
request: Request<Vec<u8, Global>>
) -> Result<Either<T, Status>, Error> where
T: DeserializeOwned,

Perform a raw HTTP request against the API and get back either an object deserialized as JSON or a Status Object.

pub async fn request_events<T>(
&'_ self,
request: Request<Vec<u8, Global>>
) -> Result<impl TryStream, Error> where
T: Clone + DeserializeOwned,

Perform a raw request and get back a stream of WatchEvent objects

impl Client

Low level discovery methods using k8s_openapi types.

Consider using the discovery module for easier-to-use variants of this functionality. The following methods might be deprecated to avoid confusion between similarly named types within discovery.

pub async fn apiserver_version(&'_ self) -> Result<Info, Error>

Returns apiserver version.

pub async fn list_api_groups(&'_ self) -> Result<APIGroupList, Error>

Lists api groups that apiserver serves.

pub async fn list_api_group_resources(
&'_ self,
apiversion: &'_ str
) -> Result<APIResourceList, Error>

Lists resources served in given API group.

Example usage:

let apigroups = client.list_api_groups().await?;
for g in apigroups.groups {
    let ver = g
        .preferred_version
        .as_ref()
        .or_else(|| g.versions.first())
        .expect("preferred or versions exists");
    let apis = client.list_api_group_resources(&ver.group_version).await?;
    dbg!(apis);
}

pub async fn list_core_api_versions(&'_ self) -> Result<APIVersions, Error>

Lists versions of core a.k.a. "" legacy API group.

pub async fn list_core_api_resources(
&'_ self,
version: &'_ str
) -> Result<APIResourceList, Error>

Lists resources served in particular core group version.