Module kube::api

API helpers for structured interaction with the Kubernetes API

Modules

entry

API helpers for get-or-create and get-and-modify patterns

Structs

Api

The generic Api abstraction

ApiResource

Information about a Kubernetes API resource

DeleteParams

Common query parameters for delete calls

DynamicObject

A dynamic representation of a kubernetes object

EvictParams

Params for evictable objects

GetParams

Common query parameters used in get calls

GroupVersionKind

Core information about an API Resource.

GroupVersionResource

Represents a type-erased object resource.

ListMeta

ListMeta describes metadata that synthetic resources must have, including lists and various status objects. A resource may have only one of {ObjectMeta, ListMeta}.

ListParams

Common query parameters used in list/delete calls on collections

LogParams

Params for logging

NotUsed

Empty struct for when data should be discarded

Object

A standard Kubernetes object with .spec and .status.

ObjectList

A generic Kubernetes object list

ObjectMeta

ObjectMeta is metadata that all persisted resources must have, which includes all objects users must create.

PartialObjectMeta

A generic representation of any object with ObjectMeta.

PatchParams

Common query parameters for patch calls

PostParams

Common query parameters for put/post calls

Preconditions

Preconditions must be fulfilled before an operation (update, delete, etc.) is carried out.

Request

A Kubernetes request builder

ScaleSpec

ScaleSpec describes the attributes of a scale subresource.

ScaleStatus

ScaleStatus represents the current status of a scale subresource.

TypeMeta

Type information that is flattened into every kubernetes object

WatchParams

Common query parameters used in watch calls on collections

Enums

Patch

Describes changes that should be applied to a resource

PropagationPolicy

Propagation policy when deleting single objects

ValidationDirective

The validation directive to use for fieldValidation when using server-side apply.

VersionMatch

Controls how the resource version parameter is applied for list calls

WatchEvent

A raw event returned from a watch query

Traits

Ephemeral

Marker trait for objects that support the ephemeral containers sub resource.

Evict

Marker trait for objects that can be evicted

Log

Marker trait for objects that has logs

PartialObjectMetaExt

Helper trait for converting ObjectMeta into useful PartialObjectMeta variants

Resource

An accessor trait for a kubernetes Resource.

ResourceExt

Helper methods for resources.