Module kube::config

Kubernetes configuration objects from ~/.kube/config or in cluster environment.

Used to populate Config that is ultimately used to construct a Client.

Unless you have issues, prefer using Config::infer and pass it to a Client.

Structs

AuthInfo

AuthInfo stores information to tell cluster who you are.

AuthProviderConfig

AuthProviderConfig stores auth for specified cloud provider.

Cluster

Cluster stores information to connect Kubernetes cluster.

Config

Configuration object detailing things like cluster URL, default namespace, root certificates, and timeouts.

Context

Context stores tuple of cluster and user information.

ExecConfig

ExecConfig stores credential-plugin configuration.

KubeConfigOptions

KubeConfigOptions stores options used when loading kubeconfig file.

Kubeconfig

Kubeconfig represents information on how to connect to a remote Kubernetes cluster

NamedAuthInfo

NamedAuthInfo associates name with authentication.

NamedCluster

NamedCluster associates name with cluster.

NamedContext

NamedContext associates name with context.

NamedExtension

NamedExtention associates name with extension.

Preferences

Preferences stores extensions for cli.