Crate kube::core

Re-exports from kube_core Crate with types and traits necessary for interacting with the Kubernetes API

This crate provides the minimal apimachinery necessary to make requests to the kubernetes API.

It does not export export a client, but it also has almost no dependencies.

Everything in this crate is re-exported from kube (even with zero features) under kube::core.

Modules

crd

Traits and tyes for CustomResources

discovery

Type information structs for API discovery

dynamic

Contains types for using resource kinds not known at compile-time.

gvk

Type information structs for dynamic resources.

metadata

Metadata structs used in traits, lists, and dynamic objects.

object

Generic object and objectlist wrappers.

params

A port of request parameter *Optionals from apimachinery/types.go

request

Request builder type for arbitrary api types

response

Generic api response types

subresource

Request builder types and parameters for subresources

util

Utils and helpers

watch

Types for the watch api

Structs

ApiResource

Information about a Kubernetes API resource

DynamicObject

A dynamic representation of a kubernetes object

ErrorResponse

An error response from the API.

GroupVersion

Core information about a family of API Resources

GroupVersionKind

Core information about an API Resource.

GroupVersionResource

Represents a type-erased object resource.

ListMeta

ListMeta describes metadata that synthetic resources must have, including lists and various status objects. A resource may have only one of {ObjectMeta, ListMeta}.

NotUsed

Empty struct for when data should be discarded

Object

A standard Kubernetes object with .spec and .status.

ObjectList

A generic Kubernetes object list

ObjectMeta

ObjectMeta is metadata that all persisted resources must have, which includes all objects users must create.

Request

A Kubernetes request builder

Status

A Kubernetes status object

TypeMeta

Type information that is flattened into every kubernetes object

Enums

Error

Core error types.

WatchEvent

A raw event returned from a watch query

Traits

CustomResourceExt

Extension trait that is implemented by kube-derive

Resource

An accessor trait for a kubernetes Resource.

ResourceExt

Helper methods for resources.

Type Definitions

Result

Convient alias for Result<T, Error>