Crate kube::runtime

Re-exports from kube-runtime Common components for building Kubernetes operators

This crate contains the core building blocks to allow users to build controllers/operators/watchers that need to synchronize/reconcile kubernetes state.

Newcomers are recommended to start with the Controller builder, which gives an opinionated starting point that should be appropriate for simple operators, but all components are designed to be usable á la carte if your operator doesn’t quite fit that mold.

Modules

controller

Runs a user-supplied reconciler function on objects when they (or related objects) are updated

events

Publishes events for objects for kubernetes >= 1.19

finalizer

Finalizer helper for Controller reconcilers

reflector

Caches objects in memory

scheduler

Delays and deduplicates Stream items

utils

Helpers for manipulating built-in streams

wait

Waits for objects to reach desired states

watcher

Watches a Kubernetes Resource for changes, with error recovery

Structs

Controller

Controller

Traits

WatchStreamExt

Extension trait for streams returned by watcher or reflector

Functions

applier

Apply a reconciler to an input stream, with a given retry policy

finalizer

Reconcile an object in a way that requires cleanup before an object can be deleted.

reflector

Caches objects from watcher::Events to a local Store

scheduler

Stream transformer that delays and deduplicates Stream items.

watcher

Watches a Kubernetes Resource for changes continuously