Trait kube::runtime::WatchStreamExt

pub trait WatchStreamExt: Stream {
    // Provided methods
    fn backoff<B>(self, b: B) -> StreamBackoff<Self, B>
       where B: Backoff,
             Self: TryStream + Sized { ... }
    fn applied_objects<K>(self) -> EventFlatten<Self, K>
       where Self: Stream<Item = Result<Event<K>, Error>> + Sized { ... }
    fn touched_objects<K>(self) -> EventFlatten<Self, K>
       where Self: Stream<Item = Result<Event<K>, Error>> + Sized { ... }
}

Extension trait for streams returned by watcher or reflector

Provided Methods

fn backoff<B>(self, b: B) -> StreamBackoff<Self, B>where B: Backoff, Self: TryStream + Sized,

Apply a Backoff policy to a Stream using StreamBackoff

fn applied_objects<K>(self) -> EventFlatten<Self, K>where Self: Stream<Item = Result<Event<K>, Error>> + Sized,

Flatten a watcher() stream into a stream of applied objects

All Added/Modified events are passed through, and critical errors bubble up.

fn touched_objects<K>(self) -> EventFlatten<Self, K>where Self: Stream<Item = Result<Event<K>, Error>> + Sized,

Flatten a watcher() stream into a stream of touched objects

All Added/Modified/Deleted events are passed through, and critical errors bubble up.